Support SUMMARY in serialization to Measurements.
This throws away quantile data in favor of the sum and count.  The sum
is reported as a gauge, and the count, not surprisingly a
counter. Further, the names of the metrics follow the typical mangling
that agentmon does.

In future versions, that otherwise might record quantiles, the space
between dotted label-pairs, probably makes sense to plop the quantile
information in. For example `some_summary__0.95.path_index`.