Fix discrepency between statsd counters, and prom counters.
agentmon uses counters in the statsd sense, that is to say:
"Since the last time we flushed, this many things happened." In more
traditional monitoring parlance, this is the derivative of a
monotonically increasing counter.

These two things are *not* compatible numerically, but a statsd style
counter is *easily* derived (no pun intended) from a monotonic counter
provided that the previous value of the counter was known.

This commit introduces a `parent` MeasurementSet, such that the previous
value can be utilized. In addition, it adds a new, non-exported field
monoCounters, such that a DerivedCounter can properly be derived.